Change build_sphinx to only collect the docs
Leave building to hatchling
